requirements.txt
tqdm
llama_index
lancedb
llama-index-embeddings-nvidia
llama-index-llms-groq
python-dotenv
